+++ proc_planner_ros2/main.py
@@ -1,11 +1,22 @@
+"""Proc Planner Main File."""
+
 import rclpy
 
-from proc_planner_ros2.ProcPlanner import ProcPlanner
+from proc_planner_ros2.proc_planner_node import ProcPlannerNode
+
 
 def main(args=None):
+    """
+    Execute package.
+
+    Keyword Arguments:
+    -----------------
+        args -- RCLPY Args (default: {None})
+
+    """
     rclpy.init(args=args)
 
-    proc_planner = ProcPlanner()
+    proc_planner = ProcPlannerNode()
 
     rclpy.spin(proc_planner)
