+++ proc_planner_ros2/proc_planner_node.py
@@ -1,0 +1,102 @@
+"""Proc Planner Node."""
+
+import rclpy
+from rcl_interfaces.msg import ParameterDescriptor
+from rclpy.node import Node
+from sonia_common_ros2.msg import PoseArray
+from geometry_msgs.msg import Pose
+from trajectory_msgs.msg import MultiDOFJointTrajectoryPoint
+from std_msgs.msg import Bool
+
+from proc_planner_ros2.trajectory_generator import TrajectoryGenerator
+
+
+class ProcPlannerNode(Node):
+    """Proc Planer Node."""
+
+    def __init__(self):
+        super().__init__("proc_planner_ros2")
+        self._sub_mult_add_pose = self.create_subscription(
+            PoseArray, "/proc_planner/send_multi_add_pose", self._mult_add_pose_cb, 10
+        )
+        self.get_logger()
+        self._sub_curr_target = self.create_subscription(Pose, "/proc_control/current_target", self._curr_target_cb, 10)
+
+        self._pub_traj = self.create_publisher(MultiDOFJointTrajectoryPoint, "/proc_planner/send_trajectory_list", 10)
+        self._pub_is_valid = self.create_publisher(Bool, "/proc_planner/is_waypoint_valid", 10)
+
+        self._latest_curr_target = None
+
+        self.declare_parameter("ts", descriptor=ParameterDescriptor(type=rclpy.Parameter.Type.DOUBLE.value))
+        self.declare_parameter("max_depth", descriptor=ParameterDescriptor(type=rclpy.Parameter.Type.DOUBLE.value))
+        self.declare_parameter(
+            "surface_warning", descriptor=ParameterDescriptor(type=rclpy.Parameter.Type.DOUBLE.value)
+        )
+        self.declare_parameter(
+            "low_speed.maximum_acceleration", descriptor=ParameterDescriptor(type=rclpy.Parameter.Type.DOUBLE.value)
+        )
+        self.declare_parameter(
+            "low_speed.maximum_velocity", descriptor=ParameterDescriptor(type=rclpy.Parameter.Type.DOUBLE.value)
+        )
+        self.declare_parameter(
+            "low_speed.maximum_angular_rate", descriptor=ParameterDescriptor(type=rclpy.Parameter.Type.DOUBLE.value)
+        )
+        self.declare_parameter(
+            "normal_speed.maximum_acceleration", descriptor=ParameterDescriptor(type=rclpy.Parameter.Type.DOUBLE.value)
+        )
+        self.declare_parameter(
+            "normal_speed.maximum_velocity", descriptor=ParameterDescriptor(type=rclpy.Parameter.Type.DOUBLE.value)
+        )
+        self.declare_parameter(
+            "normal_speed.maximum_angular_rate", descriptor=ParameterDescriptor(type=rclpy.Parameter.Type.DOUBLE.value)
+        )
+        self.declare_parameter(
+            "high_speed.maximum_acceleration", descriptor=ParameterDescriptor(type=rclpy.Parameter.Type.DOUBLE.value)
+        )
+        self.declare_parameter(
+            "high_speed.maximum_velocity", descriptor=ParameterDescriptor(type=rclpy.Parameter.Type.DOUBLE.value)
+        )
+        self.declare_parameter(
+            "high_speed.maximum_angular_rate", descriptor=ParameterDescriptor(type=rclpy.Parameter.Type.DOUBLE.value)
+        )
+        self.get_logger().info("Parameters declared")
+        self._ros_params = {}
+
+        for param in self.get_parameters([
+                "ts",
+                "max_depth",
+                "surface_warning",
+                "low_speed.maximum_acceleration",
+                "low_speed.maximum_velocity",
+                "low_speed.maximum_angular_rate",
+                "normal_speed.maximum_acceleration",
+                "normal_speed.maximum_velocity",
+                "normal_speed.maximum_angular_rate",
+                "high_speed.maximum_acceleration",
+                "high_speed.maximum_velocity",
+                "high_speed.maximum_angular_rate",
+            ]):
+            self._ros_params[param.name] = param.value
+
+        for param in self._ros_params.items():
+            self.get_logger().info(param + ": " + str(param))
+
+    def _mult_add_pose_cb(self, msg: PoseArray):
+
+        if self._latest_curr_target is None:
+            return
+        traj_gen = TrajectoryGenerator(msg, self._latest_curr_target, self._ros_params, self.get_logger())
+
+        valid_msg = Bool()
+        valid_msg.data = traj_gen.is_valid
+        self._pub_is_valid.publish(valid_msg)
+
+        if traj_gen.status == traj_gen.TrajectoryStatus.RECEIVED_VALID_WAYPTS:
+            traj_msg = traj_gen.compute()
+            if traj_msg is not None:
+                self._pub_traj.publish(traj_msg)
+
+        self._latest_curr_target = None
+
+    def _curr_target_cb(self, msg: Pose):
+        self._latest_curr_target = msg
