+++ test/test_copyright.py
